fix(pr): address coderabbit review
- release.yml: drop the leftover stash conflict markers around the
  updater env block so the workflow YAML parses again.
- HomeBanners: actually apply the tone-specific button classes so the
  danger variant renders coral instead of falling back to amber.
- HomeBanners.test: clear the openUrl mock between cases so each
  assertion only sees its own click.
- Conversations: drop the duplicate composerInteractionBlocked guard
  (already early-returned upstream).
- Conversations.test: cover the rustChat=false branch.
- Home: don't log the raw user object on every render — it carries
  ids/PII; remove the diagnostic.
- app_state/ops: split inline tests into ops_tests.rs and assert both
  the fresh and expired branches of the current-user cache TTL.
